requirements.txt
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
Werkzeug==2.3.7
zipp==3.16.2
python-dotenv==1.0.0